+++ .gitignore
@@ -10,7 +10,6 @@
 .vagrant
 .vscode/*
 /downloads
-/Gruntfile.local.js
 /solr/vendor
 /themes/minktest
 /vendor
