+++ themes/custom/scss/finna.scss
@@ -7,11 +7,11 @@
 // Finna Bootstrap variable overrides
 @import "../../finna2/scss/global/bootstrap-variable-overrides";
 
-// Font Awesome
-@import "../../bootstrap3/scss/vendor/font-awesome/font-awesome";
+// Font Awesome (loaded via include path to silence warnings from dependencies)
+@import "font-awesome/font-awesome";
 
-// Bootstrap v3.4.1
-@import "../../bootstrap3/scss/vendor/bootstrap/bootstrap";
+// Bootstrap v3.4.1 (loaded via include path to silence warnings from dependencies)
+@import "vendor/bootstrap/bootstrap";
 
 // Finna variables
 @import "../../finna2/scss/global/variables";
