+++ themes/finna2/less/finna/common.less
@@ -1,3 +1,7 @@
+/* #SCSS>
+@use 'sass:color';
+<#SCSS */
+
 .page-content {
   width: 100%;
   @media (min-width: @screen-lg) {
@@ -291,7 +295,12 @@ table#componentparts {
     }
   }
   .even td {
+/* #SCSS>
+    background: color.adjust(@nav-tabs-border-color, $alpha: 0.8);
+<#SCSS */
+/* #LESS> */
     background: fade(@nav-tabs-border-color, 5%);
+/* <#LESS */
   }
   td a {
     font-weight: bold;
