+++ themes/finna2/less/finna/facets.less
@@ -396,7 +396,7 @@ span.badge.cb-facet {
     width: 100%;
     z-index: 1050;
     position: absolute;
-    background: fade(#fff, 80%);
+    background: rgba(255, 255, 255, 0.8);
     text-align: center;
     padding: 10px;
   }
