+++ themes/finna2/less/finna/feed.less
@@ -180,7 +180,7 @@
                 flex-shrink: 0;
                 img {
                     object-fit: cover;
-                    aspect-ratio: (3 / 2);
+                    aspect-ratio: calc(3 / 2);
                     width: 150px;
                 }
                 margin-right: 10px;
