+++ themes/finna2/less/finna/mobile-toolbar.less
@@ -185,10 +185,10 @@
     &.sticky {
       box-shadow: 0 3px 3px rgba(0, 0, 0, 0.3);
     }
-    margin-left: -(@grid-gutter-width / 2);
-    margin-right: -(@grid-gutter-width / 2);
-    padding-left: (@grid-gutter-width / 2);
-    padding-right: (@grid-gutter-width / 2);
+    margin-left: -calc(@grid-gutter-width / 2);
+    margin-right: -calc(@grid-gutter-width / 2);
+    padding-left: calc(@grid-gutter-width / 2);
+    padding-right: calc(@grid-gutter-width / 2);
     padding-top: 12px;
     padding-bottom: 12px;
     z-index: 500;
