+++ themes/finna2/less/finna/organisation-info.less
@@ -650,7 +650,7 @@
                 table {
                     th,
                     td {
-                        width: (100%/3);
+                        width: calc(100%/3);
                     }
                     th:nth-of-type(2),
                     td:nth-of-type(2) {
