+++ themes/finna2/less/global/navbar-bootstrap.less
@@ -99,14 +99,14 @@
 }
 
 .navbar-nav {
-  margin: (@navbar-padding-vertical / 2) - @navbar-padding-horizontal;
+  margin: calc((@navbar-padding-vertical / 2) - @navbar-padding-horizontal);
 
   > li > a {
     padding-top: @navbar-padding-vertical;
     padding-bottom: @navbar-padding-vertical;
-    line-height: (@navbar-height / 2);
-    margin-top: (@navbar-height / 4);
-    margin-bottom: (@navbar-height / 4);
+    line-height: calc(@navbar-height / 2);
+    margin-top: calc(@navbar-height / 4);
+    margin-bottom: calc(@navbar-height / 4);
     font-size: @navbar-font-size;
     font-weight: @navbar-font-weight;
     @media (min-width:@screen-md-min) {
@@ -120,7 +120,7 @@
       border-left: solid 1px @navbar-default-link-color;
     }
     height: ((@navbar-height) - 24);
-    margin-top: ((@navbar-height - (@navbar-height - 24)) / 2);
+    margin-top: calc((@navbar-height - (@navbar-height - 24)) / 2);
   }
   &.lang-1 {
     margin-left: 10px;
@@ -201,7 +201,7 @@
     margin: 0;
 
     &.navbar-main {
-      margin-bottom: (@navbar-height / 4);
+      margin-bottom: calc(@navbar-height / 4);
     }
   }
   .btn {
