+++ themes/finna2/scss/finna.scss
@@ -3,10 +3,10 @@
 
 @import "global/bootstrap-variable-overrides";
 
-@import "../../bootstrap3/scss/vendor/font-awesome/font-awesome";
+@import "font-awesome/font-awesome";
 
 // Bootstrap v3.4.1
-@import "../../bootstrap3/scss/vendor/bootstrap/bootstrap";
+@import "bootstrap/bootstrap";
 
 // Finna variables
 @import "global/variables";
