+++ themes/finna2/scss/finna/facets.scss
@@ -396,7 +396,7 @@ span.badge.cb-facet {
     width: 100%;
     z-index: 1050;
     position: absolute;
-    background: rgba(#fff, (80/100));
+    background: rgba(255, 255, 255, 0.8);
     text-align: center;
     padding: 10px;
   }
