+++ themes/finna2/scss/finna/map.scss
@@ -37,7 +37,7 @@ div.selection-map-canvas {
       position: relative;
       padding: 5px;
       border-radius: 50%;
-      background: rgba(#fff, (80/100));
+      background: rgba(255, 255, 255, 0.8);
     }
   }
   &.leaflet-edit-resize {
@@ -47,7 +47,7 @@ div.selection-map-canvas {
       font-size: 20px;
       padding: 5px;
       border-radius: 50%;
-      background: rgba(#fff, (80/100));
+      background: rgba(255, 255, 255, 0.8);
     }
   }
 }
