+++ themes/finna2/scss/finna/organisation-info.scss
@@ -1,5 +1,5 @@
 // *** Common styles for organisation info page and widget ***
-$organisation-info-max-w: 400px !default;
+$organisation-info-max-w: 400px;
 
 .organisation-info {
     .finna-panel {
@@ -650,7 +650,7 @@ $organisation-info-max-w: 400px !default;
                 table {
                     th,
                     td {
-                        width: (100%/3);
+                        width: calc(100%/3);
                     }
                     th:nth-of-type(2),
                     td:nth-of-type(2) {
