+++ themes/finna2/scss/finna/page-selectors.scss
@@ -57,7 +57,7 @@
       border-radius: $border-radius-base;
       @media(max-width:$screen-xs-max) {
         font-size: 1em;
-        padding: ($padding-base-horizontal/1.5);
+        padding: calc(#{$padding-base-horizontal}/1.5);
       }
     }
