+++ themes/finna2/scss/finna/search.scss
@@ -1,7 +1,7 @@
 // Thumbnail sizes for media items: search results, list entries, record views
-$thumbnail-width-small:  100px !default;
-$thumbnail-width-medium: 150px !default;
-$thumbnail-width-large:  200px !default;
+$thumbnail-width-small:  100px;
+$thumbnail-width-medium: 150px;
+$thumbnail-width-large:  200px;
 
 /* Results */
 /* --- Layout --- */
